only delete old image when editing link if old image is not the same as the new one

+++ apps/web/app/api/links/bulk/route.ts
@@ -249,17 +249,22 @@ export const PATCH = withWorkspace(async ({ req, workspace, headers }) => {
       : [];
 
   waitUntil(
-    Promise.allSettled(
-      links.map(async (link) => {
-        // delete old proxy image urls if exist and match the link ID
-        if (
-          link.image &&
-          link.image.startsWith(`${R2_URL}/images/${link.id}`)
-        ) {
-          storage.delete(link.image.replace(`${R2_URL}/`, ""));
-        }
-      }),
-    ),
+    (async () => {
+      if (data.proxy && data.image) {
+        await Promise.allSettled(
+          links.map(async (link) => {
+            // delete old proxy image urls if exist and match the link ID
+            if (
+              link.image &&
+              link.image.startsWith(`${R2_URL}/images/${link.id}`) &&
+              link.image !== data.image
+            ) {
+              storage.delete(link.image.replace(`${R2_URL}/`, ""));
+            }
+          }),
+        );
+      }
+    })(),
   );
 
   return NextResponse.json([...response, ...errorLinks], { headers });
